requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.0
itsdangerous==2.1.2
Pillow==10.0.1
python-dotenv==1.0.0